#2087 rename errorMessages to messages move yupLocales to translatedMessages remove intl from WalletBrowser, use normal FormattedMessage instead
