+++ app/modules/wallet-selector/actions.ts
@@ -2,7 +2,7 @@ import { createAction, createSimpleAction } from "../actionsUtils";
 import { browserWizardActions } from "./browser-wizard/actions";
 import { ledgerWizardActions } from "./ledger-wizard/actions";
 import { lightWizardActions } from "./light-wizard/actions";
-import {ErrorWithData} from "../../components/translatedMessages/errorMessages";
+import {ErrorWithData} from "../../components/translatedMessages/messages";
 
 const actions = {
   reset: () => createSimpleAction("WALLET_SELECTOR_RESET"),
