+++ app/modules/wallet-selector/ledger-wizard/actions.ts
@@ -1,6 +1,6 @@
 import { createAction, createSimpleAction } from "../../actionsUtils";
 import { ILedgerAccount } from "./reducer";
-import {ErrorWithData} from "../../../components/translatedMessages/errorMessages";
+import {ErrorWithData} from "../../../components/translatedMessages/messages";
 
 export const LEDGER_WIZARD_SIMPLE_DERIVATION_PATHS = ["44'/60'/1'/0", "44'/60'/0'/0"]; // TODO this should be taken from config
